+++ src/app/main/package-selection/packages-page/packages-page.component.ts
@@ -6,8 +6,9 @@ import {
 import { Observable } from 'rxjs';
 import { PackageList, Package } from '../../../core/models/package.model';
 import { map, tap } from 'rxjs/operators';
-import { StepOneComponent } from '../step-one/step-one.component';
-import { StepTwoComponent } from '../step-two/step-two.component';
+import { StepOneComponent } from './step-one/step-one.component';
+import { StepTwoComponent } from './step-two/step-two.component';
+import { StepThreeComponent } from './step-three/step-three.component';
 
 @Component({
   selector: 'app-packages-page',
@@ -20,6 +21,7 @@ import { StepTwoComponent } from '../step-two/step-two.component';
 export class PackagesPageComponent {
   @ViewChild(StepOneComponent) stepOneComponent: StepOneComponent;
   @ViewChild(StepTwoComponent) stepTwoComponent: StepTwoComponent;
+  @ViewChild(StepThreeComponent) stepThreeComponent: StepThreeComponent;
 
   constructor() {}
 
@@ -29,5 +31,9 @@ export class PackagesPageComponent {
 
   get formStepTwo() {
     return this.stepTwoComponent ? this.stepTwoComponent.formStep2 : null;
+  }
+
+  get formStepThree() {
+    return this.stepThreeComponent ? this.stepThreeComponent.formStep3 : null;
   }
 }
