--- src/app/main/package-selection/step-one/step-one.component.html
@@ -1,22 +1,0 @@
-<form fxLayout="column" fxLayout.gt-md="row" [formGroup]="formStep1">
-  <button hidden matStepperNext #nextButton></button>
-  <app-package-table
-    fxFlex
-    [formGroup]="formStep1"
-    [applyChanges]="nextPage"
-  ></app-package-table>
-
-  <app-package-details
-    *ngIf="selectedPackage | async"
-    fxFlex
-  ></app-package-details>
-</form>
-<p *ngIf="formStep1.pristine">
-  foo
-</p>
-<ng-container *ngIf="!formStep1.pristine && !formStep1.valid">
-  <mat-error>Errors:</mat-error>
-  <mat-error *ngFor="let error of formStep1.errors | keyvalue">
-    {{ error.value }}
-  </mat-error>
-</ng-container>
